+++ src/common/interfaces/fps/Picked/IListInterfaces.ts
@@ -1,0 +1,8 @@
+
+import type { IPickedList } from './IPickedList'
+import type { IPickedView } from './IPickedView'
+import type { IPickedWebBasic } from './IPickedWebBasic'
+import type { IZBasicItemInfo } from './IZBasicItemInfo'
+
+export { IPickedList, IPickedView, IPickedWebBasic, IZBasicItemInfo };
+