+++ src/common/interfaces/fps/Picked/IPickedList.ts
@@ -1,0 +1,7 @@
+
+export interface IPickedList {
+  title: string;
+  name: string;
+  guid: string;
+  isLibrary: boolean;
+}