+++ src/common/interfaces/fps/Picked/IPickedView.ts
@@ -1,0 +1,6 @@
+
+export interface IPickedView {
+  title: string;
+  name: string;
+  guid: string;
+}