+++ src/common/interfaces/fps/Picked/IPickedWebBasic.ts
@@ -1,0 +1,10 @@
+
+export interface IPickedWebBasic {
+  title: string;
+  ServerRelativeUrl: string;
+  guid: string;
+  url: string;
+  siteIcon: string;
+  error?: string;
+  HasUniqueRoleAssignments?: boolean;
+}